+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-map
 1.66 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Fixed collage popup.
 
 
 1.65 (2011-04-14)